+++ src/authorizationserver/oauth2_auth.go
@@ -1,0 +1,82 @@
+package authorizationserver
+
+import (
+	"log"
+	"net/http"
+	"github.com/azukaar/cosmos-server/src/utils"
+)
+
+func authEndpoint(rw http.ResponseWriter, req *http.Request) {
+	// This context will be passed to all methods.
+	ctx := req.Context()
+	
+	
+	if utils.LoggedInOnly(rw, req) != nil {
+		return
+	}
+
+	nickname := req.Header.Get("x-cosmos-user")
+	
+	hostname := utils.GetMainConfig().HTTPConfig.Hostname
+	if utils.IsHTTPS {
+		hostname = "https://" + hostname
+	} else {
+		hostname = "http://" + hostname
+	}
+
+	// Let's create an AuthorizeRequest object!
+	// It will analyze the request and extract important information like scopes, response type and others.
+	ar, err := oauth2.NewAuthorizeRequest(ctx, req)
+	if err != nil {
+		log.Printf("Error occurred in NewAuthorizeRequest: %+v", err)
+		oauth2.WriteAuthorizeError(ctx, rw, ar, err)
+		return
+	}
+
+	// let's see what scopes the user gave consent to
+	for _, scope := range req.PostForm["scopes"] {
+		ar.GrantScope(scope)
+	}
+
+	// Now that the user is authorized, we set up a session:
+	mySessionData := newSession(nickname, req)
+
+	// When using the HMACSHA strategy you must use something that implements the HMACSessionContainer.
+	// It brings you the power of overriding the default values.
+	//
+	// mySessionData.HMACSession = &strategy.HMACSession{
+	//	AccessTokenExpiry: time.Now().Add(time.Day),
+	//	AuthorizeCodeExpiry: time.Now().Add(time.Day),
+	// }
+	//
+
+	// If you're using the JWT strategy, there's currently no distinction between access token and authorize code claims.
+	// Therefore, you both access token and authorize code will have the same "exp" claim. If this is something you
+	// need let us know on github.
+	//
+	// mySessionData.JWTClaims.ExpiresAt = time.Now().Add(time.Day)
+
+	// It's also wise to check the requested scopes, e.g.:
+	// if ar.GetRequestedScopes().Has("admin") {
+	//     http.Error(rw, "you're not allowed to do that", http.StatusForbidden)
+	//     return
+	// }
+
+	// Now we need to get a response. This is the place where the AuthorizeEndpointHandlers kick in and start processing the request.
+	// NewAuthorizeResponse is capable of running multiple response type handlers which in turn enables this library
+	// to support open id connect.
+	response, err := oauth2.NewAuthorizeResponse(ctx, ar, mySessionData)
+
+	// Catch any errors, e.g.:
+	// * unknown client
+	// * invalid redirect
+	// * ...
+	if err != nil {
+		log.Printf("Error occurred in NewAuthorizeResponse: %+v", err)
+		oauth2.WriteAuthorizeError(ctx, rw, ar, err)
+		return
+	}
+
+	// Last but not least, send the response!
+	oauth2.WriteAuthorizeResponse(ctx, rw, ar, response)
+}