+++ src/user/token.go
@@ -23,7 +23,7 @@ func RefreshUserToken(w http.ResponseWriter, req *http.Request) (utils.User, err
 	// if new install
 	if config.NewInstall {
 		// check route
-		if req.URL.Path != "/cosmos/api/status" && req.URL.Path != "/cosmos/api/newInstall" {
+		if req.URL.Path != "/cosmos/api/status" && req.URL.Path != "/cosmos/api/newInstall" && req.URL.Path != "/cosmos/api/dns" {
 			json.NewEncoder(w).Encode(map[string]interface{}{
 				"status": "NEW_INSTALL",
 			})